refactor: dto repository change
