+++ src/main/java/codesquad/web/UserController.java
@@ -2,7 +2,7 @@ package codesquad.web;
 
 import codesquad.UnAuthenticationException;
 import codesquad.domain.User;
-import codesquad.dto.LoginDTO;
+import codesquad.web.dto.LoginDTO;
 import codesquad.security.HttpSessionUtils;
 import codesquad.security.LoginUser;
 import codesquad.service.UserService;
